Remove tty flag for docker, omit sourcing common
Having the '-t' flag present for docker exec allocates a tty.  This prevents running the command on all cluster members for the salt master via a command such as

> salt '*' cmd.run 'so-redis-count'

In addition, there are no dependencies on any definitions in so-common.  There is no need to source the file.  In fact, doing so adds the "must be root" dependency, which if the user is a member of the docker group, is not true for this command.  The error when trying to run docker exec from a user that is not a member of the docker group is obviously enough a permissions error that a user should know to run it as root.