es_retention_setup
This patch restores a feature of v2.3 where elasticsearch can use up to 80% of the disk in certain situations.

If the system does not store sensor data (eg. is standalone, eval, or heave node), or if the /nsm/elasticsearch directory is its own filesystem, then more than the default 50% can be used.  This patch checks for those conditions on setup and if matches, sets the retention_pct grain at 80% in the elasticsearch pillar.